pyproject.toml
[tool.poetry]
name = "medical-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
qdrant-client = "^1.12.1"
langchain-community = "^0.3.9"
pypdf = "^5.1.0"
langchain-qdrant = "^0.2.0"
langchain-openai = "^0.2.10"
python-dotenv = "^1.0.1"
streamlit = "^1.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
